Update order queries
