+++ src/queries/orders.ts
@@ -1,5 +1,8 @@
 import gql from "graphql-tag";
 
+import { orderDetailFragment } from "../fragments/user";
+import { fragmentInvoice } from "../fragments/invoice";
+
 export const ordersByUser = gql`
   query OrdersByUser($perPage: Int!, $after: String) {
     me {
@@ -45,6 +48,28 @@ export const ordersByUser = gql`
             }
           }
         }
+      }
+    }
+  }
+`;
+
+export const orderDetailsByTokenQuery = gql`
+  ${orderDetailFragment}
+  query OrderByToken($token: UUID!) {
+    orderByToken(token: $token) {
+      ...OrderDetail
+    }
+  }
+`;
+
+export const userOrderDetailsByTokenQuery = gql`
+  ${orderDetailFragment}
+  ${fragmentInvoice}
+  query UserOrderByToken($token: UUID!) {
+    orderByToken(token: $token) {
+      ...OrderDetail
+      invoices {
+        ...InvoiceFragment
       }
     }
   }
