+++ src/queries/user.ts
@@ -1,30 +1,6 @@
 import gql from "graphql-tag";
 
 import { userFragment } from "../fragments/auth";
-import { orderDetailFragment } from "../fragments/user";
-import { fragmentInvoice } from "../fragments/invoice";
-
-export const orderDetailsByTokenQuery = gql`
-  ${orderDetailFragment}
-  query OrderByToken($token: UUID!) {
-    orderByToken(token: $token) {
-      ...OrderDetail
-    }
-  }
-`;
-
-export const userOrderDetailsByTokenQuery = gql`
-  ${orderDetailFragment}
-  ${fragmentInvoice}
-  query UserOrderByToken($token: UUID!) {
-    orderByToken(token: $token) {
-      ...OrderDetail
-      invoices {
-        ...InvoiceFragment
-      }
-    }
-  }
-`;
 
 export const getUserDetailsQuery = gql`
   ${userFragment}
